requirements.txt
pycryptodome==3.20.0
requests==2.31.0
base58==2.1.1
